Chapter02/mysql2spanner/requirements.txt
google-cloud-spanner==1.3.0
